dom/base/test/jsmodules/test_moduleParsedAsModule.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test module script parsed as module</title>
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<script>
  var wasRun = false;
  var hadSyntaxError = false;

  SimpleTest.waitForExplicitFinish();
  window.onerror = handleError;

  function handleError(message, url, line, column, error) {
    hadSyntaxError = error instanceof SyntaxError;
  }

  function testError() {
    ok(wasRun, 'Check module was run');
    ok(!hadSyntaxError, 'Check that no SyntaxError was thrown');
    SimpleTest.finish();
  }
</script>
<script type="module" src="module_testSyntax.mjs"></script>
<body onload='testError()'></body>
